% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_coeff.R
\name{get_coeff}
\alias{get_coeff}
\title{Create Coefficients Data Frame}
\usage{
get_coeff(x, model = "H1")
}
\arguments{
\item{x}{an lm model object}

\item{model}{The model number (e.g., "H1")}
}
\description{
Create Coefficients Data Frame
}
