% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_coeff.R
\name{regression_coeff}
\alias{regression_coeff}
\title{Regression Coefficients}
\usage{
regression_coeff(x, y = NULL, z = NULL, ci = FALSE, se = FALSE)
}
\arguments{
\item{x}{a model object}

\item{y}{a model object}

\item{z}{a model object}

\item{ci}{display unstandardized confidence intervals? default = FALSE}

\item{se}{display standard errors? default = FALSE}
}
\description{
Displays a table for regression coefficients.
Multiple models can be added (x, y, and z).
}
