% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_modelsig.R
\name{regression_modelsig}
\alias{regression_modelsig}
\title{Regression Model Fit}
\usage{
regression_modelsig(x, y = NULL, z = NULL, print = TRUE)
}
\arguments{
\item{x}{a model object}

\item{y}{a model object}

\item{z}{a model object}

\item{print}{Create a knitr table for displaying as html table?
(default = TRUE)}
}
\description{
Displays a table for the fit of regression models. Multiple models can
be added (x, y, and z).
}
