% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_standardized.R
\name{regression_standardized}
\alias{regression_standardized}
\title{Standardized Regression Model}
\usage{
regression_standardized(x)
}
\arguments{
\item{x}{a model object}
}
\description{
Convert an unstandardized lm model object into a standardized lm model object.
This standardizes all variables in the model and reruns the lm() on the
standardized values.
}
