% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_tables.R
\name{regression_tables}
\alias{regression_tables}
\title{Regression Model Tables}
\usage{
regression_tables(x, y = NULL, z = NULL, ci = TRUE, se = TRUE, print = TRUE)
}
\arguments{
\item{x}{a model object}

\item{y}{a model object}

\item{z}{a model object}

\item{ci}{display unstandardized confidence intervals? default = TRUE}

\item{se}{display standard errors? default = TRUE}

\item{print}{Create a knitr table for displaying as html table?
(default = TRUE)}
}
\description{
Display all tables relevant for Regression models. You can include up to
three models (x, y, and z) for hierarchical regression results.
}
